requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
contourpy==1.2.0
crispy==0.7.4
crispy-bootstrap4==2023.1
cycler==0.12.1
Django==4.2.6
django-bootstrap4==23.2
django-crispy-forms==2.1
django-filter==23.5
django-localflavor==4.0
fabio==2023.10.0
fonttools==4.44.3
h5py==3.10.0
hdf5plugin==4.3.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-stdnum==1.19
silx==1.1.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
